+++ src/lib/field.unit-test.ts
@@ -132,6 +132,13 @@ equivalent({ from: [field], to: unit })(
   (x) => x === 0n || x === 1n || throwError('not boolean'),
   (x) => x.assertBool()
 );
+equivalent({ from: [field], to: unit })(
+  (x) => x === 0n || x === 1n || throwError('not boolean'),
+  (x) => {
+    let y = Provable.witness(Field, () => x.div(2));
+    y.mul(2).assertBool();
+  }
+);
 equivalent({ from: [smallField], to: bool })(
   (x) => (x & 1n) === 0n,
   (x) => x.isEven()