+++ src/lib/gadgets/gadgets.ts
@@ -8,7 +8,7 @@ import {
   rangeCheck64,
   rangeCheck32,
   rangeCheckN,
-  isInRangeN,
+  isDefinitelyInRangeN,
   rangeCheck8,
 } from './range-check.js';
 import {
@@ -22,11 +22,15 @@ import {
   leftShift32,
 } from './bitwise.js';
 import { Field } from '../core.js';
-import { ForeignField, Field3, Sum } from './foreign-field.js';
+import {
+  ForeignField,
+  Field3,
+  Sum as ForeignFieldSum,
+} from './foreign-field.js';
 import { divMod32, addMod32 } from './arithmetic.js';
 import { SHA256 } from './sha256.js';
 
-export { Gadgets };
+export { Gadgets, Field3, ForeignFieldSum };
 
 const Gadgets = {
   /**
@@ -109,24 +113,26 @@ const Gadgets = {
   },
 
   /**
-   * Checks whether the input value is in the range [0, 2^n). `n` must be a multiple of 16.
+   * Returns a boolean which being true proves that x is in the range [0, 2^n).
    *
-   * This function proves that the provided field element can be represented with `n` bits.
-   * If the field element exceeds `n` bits, `Bool(false)` is returned and `Bool(true)` otherwise.
+   * **Beware**: The output being false does **not** prove that x is not in the range [0, 2^n).
+   * This should not be viewed as a standalone provable method but as an advanced helper function
+   * for gadgets which need a weakened form of range check.
    *
-   * @param x - The value to be range-checked.
+   * @param x - The value to be weakly range-checked.
    * @param n - The number of bits to be considered for the range check.
    *
-   * @returns a Bool indicating whether the input value is in the range [0, 2^n).
+   * @returns a Bool that is definitely only true if the input is in the range [0, 2^n),
+   * but could also be false _even if_ the input is in the range [0, 2^n).
    *
    * @example
    * ```ts
    * const x = Provable.witness(Field, () => Field(12345678n));
-   * let inRange = Gadgets.isInRangeN(32, x); // return Bool(true)
+   * let definitelyInRange = Gadgets.isDefinitelyInRangeN(32, x); // could be true or false
    * ```
    */
-  isInRangeN(n: number, x: Field) {
-    return isInRangeN(n, x);
+  isDefinitelyInRangeN(n: number, x: Field) {
+    return isDefinitelyInRangeN(n, x);
   },
   /*
    * Asserts that the input value is in the range [0, 2^16).
@@ -686,7 +692,12 @@ const Gadgets = {
      * ForeignField.assertMul(xMinusY, z, aPlusBPlusC, f);
      * ```
      */
-    assertMul(x: Field3 | Sum, y: Field3 | Sum, z: Field3 | Sum, f: bigint) {
+    assertMul(
+      x: Field3 | ForeignFieldSum,
+      y: Field3 | ForeignFieldSum,
+      z: Field3 | ForeignFieldSum,
+      f: bigint
+    ) {
       return ForeignField.assertMul(x, y, z, f);
     },
 
@@ -818,18 +829,3 @@ const Gadgets = {
    */
   SHA256: SHA256,
 };
-
-export namespace Gadgets {
-  /**
-   * A 3-tuple of Fields, representing a 3-limb bigint.
-   */
-  export type Field3 = [Field, Field, Field];
-
-  export namespace ForeignField {
-    /**
-     * Lazy sum of {@link Field3} elements, which can be used as input to {@link Gadgets.ForeignField.assertMul}.
-     */
-    export type Sum = Sum_;
-  }
-}
-type Sum_ = Sum;