+++ src/lib/proof-system/prover-keys.ts
@@ -124,7 +124,7 @@ function encodeProverKey(value: SnarkKey): Uint8Array {
     }
     default:
       value satisfies never;
-      throw Error('todo');
+      throw Error('unreachable');
   }
 }
 
@@ -163,7 +163,7 @@ function decodeProverKey(header: SnarkKeyHeader, bytes: Uint8Array): SnarkKey {
     }
     default:
       header satisfies never;
-      throw Error('todo');
+      throw Error('unreachable');
   }
 }